test: cover sign-in help disclosure
